PR: MNG-1288 Submitted by: Victor Yushenko avoid copying directory over the top of itself
git-svn-id: https://svn.apache.org/repos/asf/maven/components/trunk@327947 13f79535-47bb-0310-9956-ffa450edef68

+++ maven-plugins/maven-war-plugin/src/main/java/org/apache/maven/plugin/war/AbstractWarMojo.java
@@ -262,7 +262,7 @@ public abstract class AbstractWarMojo
 
         File webappClassesDirectory = new File( webappDirectory, WEB_INF + "/classes" );
 
-        if ( getClassesDirectory().exists() )
+        if ( getClassesDirectory().exists() && (!getClassesDirectory().equals(webappClassesDirectory)))
         {
             FileUtils.copyDirectoryStructure( getClassesDirectory(), webappClassesDirectory );
         }
